Simplify search backend configuration
