+++ saleor/search/backends/picker.py
@@ -1,6 +1,6 @@
 from __future__ import unicode_literals
 
-import importlib
+from django.utils.module_loading import import_module
 from django.conf import settings
 
 
@@ -10,9 +10,7 @@ def pick_backend():
     :rtype: callable with one argument of type str
 
     '''
-    backend = importlib.import_module(
-        settings.SEARCH_BACKENDS[settings.SEARCH_BACKEND]['storefront'])
-    return backend.search
+    return import_module(settings.SEARCH_BACKEND).search_storefront
 
 
 def pick_dashboard_backend():
@@ -21,6 +19,4 @@ def pick_dashboard_backend():
     :rtype: callable with one argument of type str
 
     '''
-    backend = importlib.import_module(
-        settings.SEARCH_BACKENDS[settings.SEARCH_BACKEND]['dashboard'])
-    return backend.search
+    return import_module(settings.SEARCH_BACKEND).search_dashboard
