+++ saleor/settings.py
@@ -366,21 +366,10 @@ ES_URL = (os.environ.get('ELASTICSEARCH_URL') or
 
 ENABLE_SEARCH = bool(ES_URL) or DB_SEARCH_ENABLED  # global search disabling
 
-SEARCH_BACKENDS = {
-    'default': {
-        'storefront': 'saleor.search.backends.postgresql',
-        'dashboard': 'saleor.search.backends.postgresql_dashboard'
-    },
-    'elasticsearch': {
-        'storefront': 'saleor.search.backends.elasticsearch',
-        'dashboard': 'saleor.search.backends.elasticsearch_dashboard'
-    }
-}
-
-SEARCH_BACKEND = 'default'
+SEARCH_BACKEND = 'saleor.search.backends.postgresql'
 
 if ES_URL:
-    SEARCH_BACKEND = 'elasticsearch'
+    SEARCH_BACKEND = 'saleor.search.backends.elasticsearch'
     INSTALLED_APPS.append('django_elasticsearch_dsl')
     ELASTICSEARCH_DSL = {
         'default': {
